Make "socials" and "projects" fields optional for member JSON
Before, if you left out the "socials" or "projects" fields, it would give an error.
If you wanted to have no socials or no projects, you had to give it an empty array.
Now, you can omit any of the fields, which I think is friendlier to members adding their own pages.